Moved Phaser-specific game files to 'Game' directory
