+++ src/Phaser/Game/gameSettings.js
